src/core/rag_system.py
from src.agents.database_agent import DatabaseAgent
from src.agents.africas_talking_agent import AfricasTalkingAgent

class RAGSystem:
    def __init__(self):
        self.agents = [
            DatabaseAgent(),
            AfricasTalkingAgent()
        ]

    def retrieve_and_generate(self, query):
        results = []
        for agent in self.agents:
            result = agent.generate_response(query)
            results.append(result)
        # Combine results from all agents
        combined_result = self.combine_results(results)
        return combined_result

    def combine_results(self, results):
        # Implement a method to combine results from multiple agents
        return results
